# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: API сервиса маркетплейса (Homework 2)
...
paths:
  /products:
    post:
      tags:
        - Products
      summary: Создание товара
      description: 'Доступно для ролей: SELLER, ADMIN'
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'

    get:
      tags:
        - Products
      summary: Список товаров с пагинацией и фильтрацией
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: Номер страницы, начиная с 0
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: status
          in: query
          description: Фильтр по статусу
          required: false
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          description: Фильтр по категории (точное совпадение)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ со списком товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductResponse'
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Products
      summary: Получение товара по ID
      operationId: getProductById
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Products
      summary: Обновление товара
      description: 'Доступно для ролей: SELLER, ADMIN'
      operationId: updateProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Products
      summary: Мягкое удаление товара
      description: 'Доступно для ролей: SELLER, ADMIN'
      operationId: deleteProduct
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Товар переведен в статус ARCHIVED
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  schemas:
...
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]
    ProductCreate:
      type: object
      required:
        - name
        - price
        - stock
        - category
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          exclusiveMinimum: 0
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          exclusiveMinimum: 0
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - category
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaginatedProductResponse:
      type: object
      required:
        - items
        - totalElements
        - page
        - size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
        page:
          type: integer
        size:
          type: integer

    # --- Orders ---
...
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    NotFoundError:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            product:
              value:
                error_code: PRODUCT_NOT_FOUND
                message: Товар не найден по ID
            order:
              value:
                error_code: ORDER_NOT_FOUND
                message: Заказ не найден по ID
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: Ошибка валидации данных
            details:
              price: "Значение должно быть больше 0"
    AccessDeniedError:
      description: Доступ запрещен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ACCESS_DENIED
            message: Недостаточно прав для выполнения операции
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT